requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-marshmallow==0.11.0
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-datastore==1.10.0
google-cloud-ndb==1.0.1
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.4.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
redis==3.4.1
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
